fixed some issues
