+++ dependent-mapping/src/main/java/com/iluwatar/dependentmapping/AlbumMapper.java
@@ -24,11 +24,11 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package com.iluwatar.denpendentmapping;
+package com.iluwatar.dependentmapping;
 
-import com.iluwatar.denpendentmapping.structure.DependentObj;
-import com.iluwatar.denpendentmapping.structure.Mapper;
-import com.iluwatar.denpendentmapping.structure.MasterObj;
+import com.iluwatar.dependentmapping.structure.DependentObj;
+import com.iluwatar.dependentmapping.structure.Mapper;
+import com.iluwatar.dependentmapping.structure.MasterObj;
 import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.sql.SQLException;
