+++ dependent-mapping/src/main/java/com/iluwatar/dependentmapping/structure/DependentObj.java
@@ -24,48 +24,36 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package com.iluwatar.denpendentmapping.structure;
-
-import java.sql.SQLException;
+package com.iluwatar.dependentmapping.structure;
 
 /**
- * Mapper which depend on the master class and dependent class,
- * and connect them with the database.
+ * A class of which instance will be depend on the
+ * MasterObj class instance.
  */
-public interface Mapper {
+public abstract class DependentObj {
+  /**
+   * title.
+   */
+  private final String title;
 
   /**
-   * Show the data from the database.
+   * Construction method.
    *
-   * @return the data in database.
+   * @param newTitle title.
    */
-  String findstatement();
+  public DependentObj(final String newTitle) {
+
+    this.title = newTitle;
+
+  }
 
   /**
-   * Updata the specific master instance and its dependent instances.
+   * get title of it.
    *
-   * @param arg the master instance.
-   * @throws SQLException the exception of SQL
+   * @return title.
    */
-  void update(MasterObj arg) throws SQLException;
+  public String getTitle() {
 
-  /**
-   * Updata the dependent instances of the specific master instance.
-   *
-   * @param arg the master instance.
-   * @throws SQLException the exception of SQL
-   */
-  void updateDepObjs(MasterObj arg) throws SQLException;
-
-  /**
-   * Insert specific a dependent instance into a master instance in database.
-   *
-   * @param dependentObj the specific dependent instance.
-   * @param seq          the sequence.
-   * @param masterObj    the specific master instance
-   * @throws SQLException the exception of SQL
-   */
-  void insertDepObj(DependentObj dependentObj,
-                    int seq, MasterObj masterObj) throws SQLException;
-
+    return title;
+  }
 }
